refactor: Extract class tftp::settings_val to new files tftpSettings.* (u-t OK!)
Running 7 test cases...
long time checks [...................................]
Summary checks 1834041
*** No errors detected
