+++ source/tftp_common.h
@@ -18,7 +18,6 @@
 #include <cassert>
 #include <cxxabi.h> // for current class/type name
 #include <functional>
-#include <memory>
 #include <ostream>
 #include <string>
 #include <syslog.h>
@@ -26,6 +25,7 @@
 #include <netinet/in.h>
 #include <unistd.h>
 #include <iostream>
+#include <memory>
 
 
 #include <iomanip>
@@ -116,56 +116,6 @@ struct is_container<
 
 template<class T>
 constexpr bool is_container_v = is_container<T>::value;
-
-// ----------------------------------------------------------------------------------
-
-/**
- * \brief Settings storage class
- *
- *  Class for store server settings.
- *  Can't simple construct, create only from settings_val::create() as shared pointer
- */
-
-class settings_val: public std::enable_shared_from_this<settings_val>
-{
-protected:
-
-  /// No public constructor
-  settings_val();
-
-public:
-
-  /// Public settings creator
-  static std::shared_ptr<settings_val> create() { return std::make_shared<settings_val>(settings_val{}); };
-
-  /// Destructor
-  virtual ~settings_val() {};
-
-  bool is_daemon; ///< Flag showing run as daemon
-
-  buffer_t local_base_; ///< Listening server address:port (sockaddr_in*)
-
-  // fb lib settings
-  std::string lib_dir;  ///< System library directory (Ubuntu 18.04: /usr/lib/x86_64-linux-gnu/)
-  std::string lib_name; ///< Firebird library filename (libfbclient.so)
-
-  // storage directory
-  std::string root_dir;                 ///< Root directory of simple TFTP server
-  std::vector<std::string> backup_dirs; ///< Search directories of simple TFTP server
-
-  // storage firebird
-  std::string db;      ///< Database of firebird
-  std::string user;    ///< User name access firebird
-  std::string pass;    ///< Password access firebird
-  std::string role;    ///< Role access firebird
-  uint16_t    dialect; ///< Firebird server dialect
-
-  //logger
-  int use_syslog;    ///< Syslog pass level logging message
-  f_log_msg_t log_;  ///< External callback for logging message
-};
-
-// ----------------------------------------------------------------------------------
 
 
 //=================================================================================================================================
